+++ HPXMLtoOpenStudio/measure.rb
@@ -208,6 +208,7 @@ class OSModel
     create_or_get_space(model, spaces, HPXML::LocationLivingSpace)
     set_foundation_and_walls_top()
     set_heating_and_cooling_seasons()
+    set_summer_and_winter_seasons(model)
     add_setpoints(runner, model, weather, spaces)
 
     # Geometry/Envelope
@@ -216,7 +217,6 @@ class OSModel
     add_rim_joists(runner, model, spaces)
     add_floors(runner, model, spaces)
     add_foundation_walls_slabs(runner, model, weather, spaces)
-    add_shading_schedule(model, weather)
     add_windows(model, spaces)
     add_doors(model, spaces)
     add_skylights(model, spaces)
@@ -1090,15 +1090,35 @@ class OSModel
     end
   end
 
-  def self.add_shading_schedule(model, weather)
-    # Use BAHSP cooling season, and not year-round or user-specified cooling season, to ensure windows use appropriate interior shading factors
-    _default_heating_months, @default_cooling_months = HVAC.get_default_heating_and_cooling_seasons(weather)
+  def self.set_summer_and_winter_seasons(model)
+    summer_start_day_num = Schedule.get_day_num_from_month_day(@hpxml.header.sim_calendar_year,
+                                                               @hpxml.header.seasons_summer_begin_month,
+                                                               @hpxml.header.seasons_summer_begin_day)
+    summer_end_day_num = Schedule.get_day_num_from_month_day(@hpxml.header.sim_calendar_year,
+                                                             @hpxml.header.seasons_summer_end_month,
+                                                             @hpxml.header.seasons_summer_end_day)
 
-    # Create cooling season schedule
-    clg_season_sch = MonthWeekdayWeekendSchedule.new(model, 'cooling season schedule', Array.new(24, 1), Array.new(24, 1), @default_cooling_months, Constants.ScheduleTypeLimitsFraction)
-    @clg_ssn_sensor = OpenStudio::Model::EnergyManagementSystemSensor.new(model, 'Schedule Value')
-    @clg_ssn_sensor.setName('cool_season')
-    @clg_ssn_sensor.setKeyName(clg_season_sch.schedule.name.to_s)
+    # 365-element array of 24-element arrays of 0 or 1
+    @summer_values = []
+    for day in 1..Constants.NumDaysInYear(@hpxml.header.sim_calendar_year)
+      if summer_end_day_num > summer_start_day_num # northern hemisphere
+        is_summer = (day >= summer_start_day_num && day <= summer_end_day_num)
+      else # southern hemisphere
+        is_summer = (day >= summer_start_day_num || day <= summer_end_day_num)
+      end
+
+      if is_summer
+        @summer_values << [1] * 24
+      else
+        @summer_values << [0] * 24
+      end
+    end
+
+    # Create summer schedule
+    summer_sch = HourlyByDaySchedule.new(model, 'summer season schedule', @summer_values, @summer_values, Constants.ScheduleTypeLimitsFraction, false)
+    @summer_season_sensor = OpenStudio::Model::EnergyManagementSystemSensor.new(model, 'Schedule Value')
+    @summer_season_sensor.setName('cool_season')
+    @summer_season_sensor.setKeyName(summer_sch.schedule.name.to_s)
   end
 
   def self.add_windows(model, spaces)
@@ -1166,7 +1186,7 @@ class OSModel
         # Apply interior/exterior shading (as needed)
         shading_vertices = Geometry.create_wall_vertices(window_length, window_height, z_origin, window.azimuth)
         shading_group = Constructions.apply_window_skylight_shading(model, window, i, shading_vertices, surface, sub_surface, shading_group,
-                                                                    shading_schedules, shading_ems, Constants.ObjectNameWindowShade, @default_cooling_months)
+                                                                    shading_schedules, shading_ems, Constants.ObjectNameWindowShade, @summer_values)
       else
         # Window is on an interior surface, which E+ does not allow. Model
         # as a door instead so that we can get the appropriate conduction
@@ -1244,7 +1264,7 @@ class OSModel
       # Apply interior/exterior shading (as needed)
       shading_vertices = Geometry.create_roof_vertices(length, width, z_origin, skylight.azimuth, tilt)
       shading_group = Constructions.apply_window_skylight_shading(model, skylight, i, shading_vertices, surface, sub_surface, shading_group,
-                                                                  shading_schedules, shading_ems, Constants.ObjectNameSkylightShade, @default_cooling_months)
+                                                                  shading_schedules, shading_ems, Constants.ObjectNameSkylightShade, @summer_values)
     end
 
     apply_adiabatic_construction(model, surfaces, 'roof')
@@ -1736,7 +1756,7 @@ class OSModel
     end
 
     Airflow.apply(model, runner, weather, spaces, @hpxml, @cfa, @nbeds,
-                  @ncfl_ag, duct_systems, airloop_map, @clg_ssn_sensor, @eri_version,
+                  @ncfl_ag, duct_systems, airloop_map, @summer_season_sensor, @eri_version,
                   @frac_windows_operable, @apply_ashrae140_assumptions, @schedules_file, @hpxml.header.vacancy_periods)
   end
 
@@ -2371,7 +2391,7 @@ class OSModel
     program.addLine('  Set htg_mode = 1')
     program.addLine("ElseIf (#{liv_load_sensors[:clg].name} > 0)") # Assign hour to cooling if cooling load
     program.addLine('  Set clg_mode = 1')
-    program.addLine("ElseIf (#{@clg_ssn_sensor.name} > 0)") # No load, assign hour to cooling if in cooling season definition (Note: natural ventilation & whole house fan only operate during the cooling season)
+    program.addLine("ElseIf (#{@summer_season_sensor.name} > 0)") # No load, assign hour to cooling if in summer season definition (Note: natural ventilation & whole house fan only operate during the summer season)
     program.addLine('  Set clg_mode = 1')
     program.addLine('Else') # No load, assign hour to heating if not in cooling season definition
     program.addLine('  Set htg_mode = 1')