+++ HPXMLtoOpenStudio/resources/hpxml_defaults.rb
@@ -23,7 +23,7 @@ class HPXMLDefaults
       has_fuel[fuel] = hpxml.has_fuel(fuel, hpxml_doc)
     end
 
-    apply_header(hpxml, epw_file)
+    apply_header(hpxml, epw_file, weather)
     apply_emissions_scenarios(hpxml, has_fuel)
     apply_utility_bill_scenarios(runner, hpxml, has_fuel)
     apply_site(hpxml)
@@ -106,7 +106,7 @@ class HPXMLDefaults
 
   private
 
-  def self.apply_header(hpxml, epw_file)
+  def self.apply_header(hpxml, epw_file, weather)
     if hpxml.header.occupancy_calculation_type.nil?
       hpxml.header.occupancy_calculation_type = HPXML::OccupancyCalculationTypeAsset
       hpxml.header.occupancy_calculation_type_isdefaulted = true
@@ -204,6 +204,23 @@ class HPXMLDefaults
     if hpxml.header.natvent_days_per_week.nil?
       hpxml.header.natvent_days_per_week = 3
       hpxml.header.natvent_days_per_week_isdefaulted = true
+    end
+
+    if hpxml.header.seasons_summer_begin_month.nil? || hpxml.header.seasons_summer_begin_day.nil? ||
+       hpxml.header.seasons_summer_end_month.nil? || hpxml.header.seasons_summer_end_day.nil?
+      if weather.header.Latitude < 0 # Southern hemisphere
+        hpxml.header.seasons_summer_begin_month = 10 # Oct 15
+        hpxml.header.seasons_summer_end_month = 4 # Apr 14
+      else # Northern hemisphere
+        hpxml.header.seasons_summer_begin_month = 4 # Apr 15
+        hpxml.header.seasons_summer_end_month = 10 # Oct 14
+      end
+      hpxml.header.seasons_summer_begin_day = 15
+      hpxml.header.seasons_summer_end_day = 14
+      hpxml.header.seasons_summer_begin_month_isdefaulted = true
+      hpxml.header.seasons_summer_begin_day_isdefaulted = true
+      hpxml.header.seasons_summer_end_month_isdefaulted = true
+      hpxml.header.seasons_summer_end_day_isdefaulted = true
     end
   end